requirements.txt
# ============================================
# LLM增强的3D高斯场景理解系统 - 依赖包
# ============================================

# 原有3DGS依赖（从environment.yml继承）
# torch>=2.0.0
# torchvision
# tqdm
# plyfile
# opencv-python

# YOLO目标检测
ultralytics>=8.0.0
supervision>=0.18.0

# 深度学习和计算机视觉
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 3D处理
open3d>=0.17.0
trimesh>=3.23.0
scikit-learn>=1.3.0

# LLM集成
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0

# 数据处理
pandas>=2.0.0
scipy>=1.11.0
pyyaml>=6.0

# 可视化和报告生成
plotly>=5.17.0
jinja2>=3.1.2

# 工具
loguru>=0.7.0

